requirements.txt
aiofiles==0.4.0
aiohttp==3.6.2
aiosqlite3==0.3.0
altgraph==0.16.1
aiosqlite3==0.3.0
astroid==2.2.5
async-timeout==3.0.1
asyncio==3.4.3
attrs==19.1.0
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.1
discord.py==1.4.1
feedparser==5.2.1
future==0.17.1
idna==2.8
isort==4.3.20
lazy-object-proxy==1.4.1
mccabe==0.6.1
multidict==4.5.2
numpy==1.15.0
pefile==2019.4.18
pipenv==2018.11.26
psutil==5.6.3
pycparser==2.19
PyInstaller==3.5
pylint==2.3.1
PyNaCl==1.3.0
pywin32-ctypes==0.2.0
requests==2.22.0
six==1.12.0
typed-ast==1.4.0
virtualenv==16.7.3
virtualenv-clone==0.5.3
websockets==6.0
WMI==1.4.9
wrapt==1.11.2
yarl==1.3.0
dropbox==9.4.0
lyricsgenius==1.8.2
gmail==0.6.3
googletrans==2.4.0
youtube_search==0.1.3
mcstatus==2.3.0
dnspython==1.15.0
dnspython3==1.15.0
wikipedia==1.4.0
Pillow==7.0.0
dblpy==0.3.3
mysqlclient==1.4.6

